Move command locking and prioritization into the protocol handler
